fix(restore): restore REST API pre-check for CVR creation (#38)
